exmdb_provider: fixup unusual argument type for common_util_get_cid_string_length
Variables to hold CIDs are uint64_t everywhere I can see; having
a 32-bit CID variable appears out of place. (However, there
may be 32-bit limits regarding content files elsewhere..)

+++ exch/exmdb_provider/common_util.cpp
@@ -5681,7 +5681,7 @@ BOOL common_util_indexing_sub_contents(
  * counter; the caller must ensure this function only called for
  * blobs previously so encoded.
  */
-static uint32_t common_util_get_cid_string_length(uint32_t cid)
+static uint32_t common_util_get_cid_string_length(uint64_t cid)
 {
 	char path[256];
 	struct stat node_stat;
